Update craptorio_main.lua, Added belt drag-locking
Moved mouse event dispatching from TIC to it's own function. Implemented belt 'drag-locking' a feature from factorio that when placing down belts, and you start drag-placing a length of belts, the direction/axis is temporarily locked as to prevent unwanted placement of belts 'off of the line', and forcing belts to be placed in a straight line.